Refactoring
- end readable when end time is reached instead for checking in an interval if bytes are written
- adjusted silent padding (subtract added chunk size)
- remove listener if readable is destroyed
